requirements.txt
pandas
weaviate-client
openai
langchain
fastapi
uvicorn
torch
transformers
langchain_openai
